requirements.txt
PyYAML~=6.0.1
urwid~=2.6.15